hust-se-backend/requirements.txt
alembic==1.17.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
Flask==2.3.3
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pycparser==2.23
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.15.0
Werkzeug==3.1.3
